% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_predict_tte}
\alias{example_predict_tte}
\title{Run a Minimal End-to-End Example}
\usage{
example_predict_tte(nsim = 200)
}
\arguments{
\item{nsim}{Number of simulations.}
}
\value{
A `PredRes` object.
}
\description{
Demonstrates how to use packaged ADTTE input with `predict_tte()`.
}
