% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_tte_models}
\alias{fit_tte_models}
\title{Fit Enrollment, Event, and Dropout Models}
\usage{
fit_tte_models(
  data,
  enrollment_model = "exponential",
  event_model = "exponential",
  dropout_model = "exponential",
  covariates = NULL
)
}
\arguments{
\item{data}{ADTTE data with required columns.}

\item{enrollment_model}{Distribution for enrollment time.}

\item{event_model}{Distribution for event time.}

\item{dropout_model}{Distribution for dropout time.}

\item{covariates}{Optional covariates for event/dropout models.}
}
\value{
A list of fitted `survreg` models.
}
\description{
Fit Enrollment, Event, and Dropout Models
}
