\name{example_predict_tte}
\alias{example_predict_tte}
\title{Run a Minimal End-to-End Example}
\usage{
example_predict_tte(nsim = 200)
}
\arguments{
  \item{nsim}{Number of simulations.}
}
\value{
A \code{predictTTE_result} object.
}
\description{
Runs a minimal workflow using packaged example data.
}
