% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predres_methods.R
\name{showDefault}
\alias{showDefault}
\title{Show Default Prediction Summary}
\usage{
showDefault(x, ...)
}
\arguments{
\item{x}{A `PredRes` object from `predict_tte()`.}

\item{...}{Unused.}
}
\description{
Prints formatted prediction summaries for a `PredRes` object.
}
